server/requirements.txt
python-dotenv
imageio-ffmpeg
imageio
openai
python-dotenv
google-cloud-texttospeech
azure-cognitiveservices-speech
algoliasearch
gunicorn
flask
flask-cors
transformers
sentencepiece
algoliasearch
pyht